+++ packages/eslint-config/index.js
@@ -39,5 +39,9 @@ module.exports = {
       'error',
       { extensions: ['.js', '.jsx', 'tsx', 'ts'] },
     ],
+    'no-use-before-define': 'off',
+    '@typescript-eslint/no-use-before-define': ['error'],
+    'no-shadow': 'off',
+    '@typescript-eslint/no-shadow': ['error'],
   },
 };