+++ packages/stories/src/components/Editor/index.stories.tsx
@@ -14,15 +14,19 @@ const EditorStory = (): ReactElement => {
     EditorState.createEmpty(),
   );
 
+  const handleOnChange = (state: EditorState) => {
+    setEditorState(state);
+  };
+
   return (
     <Editor
       editorState={editorState}
-      onChange={setEditorState}
+      onChange={handleOnChange}
       placeholder={text('placeholder', 'placeholder')}
     />
   );
 };
 
-export const mango = Mango<{}>(EditorStory);
+export const mango = Mango(EditorStory);
 
-export const skyline = Skyline<{}>(EditorStory);
+export const skyline = Skyline(EditorStory);