+++ packages/eslint-config/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.17](https://github.com/MangoYellowH/mango-toolkit/compare/@mango-solution/eslint-config@0.0.16...@mango-solution/eslint-config@0.0.17) (2021-07-10)
+
+**Note:** Version bump only for package @mango-solution/eslint-config
+
+
+
+
+
+## [0.0.16](https://github.com/MangoYellowH/mango-toolkit/compare/@mango-solution/eslint-config@0.0.15...@mango-solution/eslint-config@0.0.16) (2021-07-10)
+
+**Note:** Version bump only for package @mango-solution/eslint-config
+
+
+
+
+
+## [0.0.15](https://github.com/MangoYellowH/mango-toolkit/compare/@mango-solution/eslint-config@0.0.14...@mango-solution/eslint-config@0.0.15) (2021-07-10)
+
+**Note:** Version bump only for package @mango-solution/eslint-config
+
+
+
+
+
 ## [0.0.14](https://github.com/MangoYellowH/mango-toolkit/compare/@mango-solution/eslint-config@0.0.13...@mango-solution/eslint-config@0.0.14) (2021-07-10)
 
 **Note:** Version bump only for package @mango-solution/eslint-config