+++ packages/mango-ui/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.50](https://github.com/MangoYellowH/mango-toolkit/compare/@mango-solution/mango-ui@0.0.49...@mango-solution/mango-ui@0.0.50) (2021-07-10)
+
+**Note:** Version bump only for package @mango-solution/mango-ui
+
+
+
+
+
+## [0.0.49](https://github.com/MangoYellowH/mango-toolkit/compare/@mango-solution/mango-ui@0.0.48...@mango-solution/mango-ui@0.0.49) (2021-07-10)
+
+**Note:** Version bump only for package @mango-solution/mango-ui
+
+
+
+
+
+## [0.0.48](https://github.com/MangoYellowH/mango-toolkit/compare/@mango-solution/mango-ui@0.0.47...@mango-solution/mango-ui@0.0.48) (2021-07-10)
+
+**Note:** Version bump only for package @mango-solution/mango-ui
+
+
+
+
+
+## [0.0.47](https://github.com/MangoYellowH/mango-toolkit/compare/@mango-solution/mango-ui@0.0.46...@mango-solution/mango-ui@0.0.47) (2021-07-10)
+
+**Note:** Version bump only for package @mango-solution/mango-ui
+
+
+
+
+
 ## [0.0.46](https://github.com/MangoYellowH/mango-toolkit/compare/@mango-solution/mango-ui@0.0.45...@mango-solution/mango-ui@0.0.46) (2021-07-10)
 
 **Note:** Version bump only for package @mango-solution/mango-ui