+++ packages/stories/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.46](https://github.com/MangoYellowH/mango-toolkit/compare/@mango-solution/stories@0.0.45...@mango-solution/stories@0.0.46) (2021-07-10)
+
+**Note:** Version bump only for package @mango-solution/stories
+
+
+
+
+
+## [0.0.45](https://github.com/MangoYellowH/mango-toolkit/compare/@mango-solution/stories@0.0.44...@mango-solution/stories@0.0.45) (2021-07-10)
+
+**Note:** Version bump only for package @mango-solution/stories
+
+
+
+
+
+## [0.0.44](https://github.com/MangoYellowH/mango-toolkit/compare/@mango-solution/stories@0.0.43...@mango-solution/stories@0.0.44) (2021-07-10)
+
+**Note:** Version bump only for package @mango-solution/stories
+
+
+
+
+
+## [0.0.43](https://github.com/MangoYellowH/mango-toolkit/compare/@mango-solution/stories@0.0.42...@mango-solution/stories@0.0.43) (2021-07-10)
+
+**Note:** Version bump only for package @mango-solution/stories
+
+
+
+
+
 ## [0.0.42](https://github.com/MangoYellowH/mango-toolkit/compare/@mango-solution/stories@0.0.41...@mango-solution/stories@0.0.42) (2021-07-10)
 
 **Note:** Version bump only for package @mango-solution/stories